Change retry behavior
The previous retry period (1 sec) is not always long enough to load
replies, so make it 10 sec instead. We still have the issue where
the replies have loaded but there are missing replies because one or
more posts are by a blocked user. When this happens, there will be no
clear indication that reply loading is done, so we need to detect this
case.

For now, the solution is bailing out after 100 msec once at least one
reply has been loaded. Context gets loaded from a single JSON file,
and it should not be common for page JS to take more than 100 msec to
fully process this file.