Send access token with context requests
The Mastodon client sends this authorization token with requests for
context. It's possible this is used to determine what posts the user has
permission to view. In any case, the result without the token (even with
logged in cookies) is sometimes different than the result with the
token, and indentation for posts that are visible to Mastodon but not to
the script was ineffective.

Bump to 2.1.5.