Fix off-by-one in Mastodon script
I think at some point in the past, the OP post was included in the
ascendents, and we subtracted one here for that reason. This seems
to not (or no longer) be the case, so e.g. for a post with zero
parent posts, the comparison here might be 0 > -1, and activate the
condition.