Fix issues with some replies appearing in the wrong order
This appears to be the result of a subtle bug in Mastodon. In certain
cases, the backend can be fooled into placing top level replies ahead
of certain replies to lower level OP direct-replies. This might be
caused by some ID based ordering mixup, or maybe that's just a symtom,
indicating e.g. when the posts arrive to the server.

Regardless, the hacky, but more efficient previous implementation was
not cutting it, because of posts that would appear with the correct
indentation in the middle of an unrelated thread. A rare issue, but
worth fixing.

Here we actually simplify the approach by just building the entire
tree and sorting children by the post time. (There is probably some
concern that AP allows lying about the post itme, but that's a rare
issue to encounter.) Then we walk down the whole list of post
elements, moving the next post to be adjacent to it as necessary.

Also, clean up some old-style use of objects as Maps.